Convert Things I Refuse to Work With section to Q&A From Myself
Shorten the description on Windows, the reader should be aware of
what the hell it is.

Remove Ubuntu from the list, although I still don't like it. The
reason is, people doing Linux should support each other, and try to
make a friendly community. They share many in common, just like
Sunni and Shia.